+++ gatsby-node.js
@@ -25,7 +25,7 @@ exports.createPages = ({ graphql, actions }) => {
 
     posts.forEach((post, index) => {
       createPage({
-        path: post.node.slug,
+        path: `/blog/${post.node.slug}`,
         component: blogPost,
         context: {
           slug: post.node.slug,
